IB/hfi: Notify IB layer of transition to active and client reregister properly
This patch removes the port active notification for the IB layer from the init
path and does it when the link goes active. It also removes the delayed client
reregister that was checked in as part of cdfa107 and does the re-register in
the mad path as it was done before.

INTERNAL: PR 126778 - WFR driver needs to not issue IB_CLIENT_REREGISTER event
INTERNAL: 	      any time the port goes ARMED to ACTIVE.

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: Brian Welty <brian.welty@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Change-Id: I180ae4df3f18cf6089fd1095651c948c412ee969