IB/hfi: Optimizations for SDMA and Expected Receives
This commit adds the following performance
optimizations in the areas of user SDMA and expected
receives:

SDMA:
-----
* Eliminate the user SDMA lock protecting the list of
  TX requests made, pending completion. We don't need
  to remember the reqs since the SDMA middle layer
  guarantees that the completion callback will be called
  for each successfully submitted request.
* Convert the request submission call from individual
  submission to a list-based submission. This simplifies
  the submission code path by eliminating the need for
  rolling any accounting back in case of submission
  failure.
* Change the progress detection routine so the descriptor
  availability check is done against the actual list
  rather then inferred by the fact that the lock sequence
  has moved.
* Force the process to sleep if the there is no room in
  the descriptor queue after one retry.
* Clean up code and remove unusesd flags.

Expected Receives:
------------------
The function which did the work of programming
the HW with pinned pages for an expected receive
took the context lock in the beginning of the
function hold it until the end. This introduced
serialization when shared contexts were used.

However, the driver does not need the lock  all
that time. It only needs it to find and reserve
the group bits that it is going to use. After that
the lock can be released since no other processes
will be using the reserved groups.

INTERNAL: PR 127453 - system crash with 512 ranks, 256 per node

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: Ia782b60b4ca01d838a23adb2c74d047ceba89993