IB/hfi: Report the correct LID prior to it being configured
By default (STL Spec Vol. 1, Table 4-1), the LID that is
reported by the driver prior to the FM setting the LID for
the node should be 0 not the Permissive LID (0xffff).

This commit fixes the driver behavior to comply with the STL
Architecture Spec.

INTERNAL: PR 127033 - ibv_devinfo reports the wrong port lid
INTERNAL: for hfi0

Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I0b17322dd7668fc72c92748a1f06eea6fa176647