IB/hfi: Handle eager buffer sizes less than 1MB
When sanitizing the value of the eager_buffer_size
module parameter, the driver rounds the value up to
the next power of two and then fits it within the
MIN and MAX theoretical limit for eager buffers.

This does not work under the following conditions:
1. the value of the module parameter is 1MB and
smaller
2. multiple-packets-per-eager-buffer is enabled
3. the eager buffer allocation chunk is 256KB
due to the requirement that at least 8 RcvArray entries
are used (1MB / 256KB = 4).

To handle eager buffer sizes of 1MB and smaller, this
commit implements the following changes:
1. the value of the eager_buffer_size parameter is
limited to be, at a minimum, the theoretical
limit of a eager RcvArray entry multiplied by 8.
2. the value of the contexts eager buffer memory'
is limited to at least 8 times the MTU (rounded
up to the next power of 2), if it less than that
3. the value of the RcvArray entry allocation chunk
is the maximum of the MTU (rounded up to the next
power of two) or the requested eager buffer size
divided by 8 (rounded down to the next power of
two)

INTERNAL: PR 127487 - Setting eager_buffer_size <= 1M
INTERNAL: (1048576) hangs system

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I77fc5f248c6bdfc17cfcd9b6ec5be1489ec4a0d3