IB/hfi: set (SM_TrapQP, SA_QP) at driver load, link up, and in Set(PortInfo)
STL specifies that the power up default, and link up default values of
SM_TrapQP, and SA_QP are 0, and 1, respectively. Set those values when
the driver is loaded, and when the link comes up. Also, allow the SM to
change these values with a Set(PortInfo).

INTERNAL: 127271 - SMA should set SM_TrapQP and SA_QP

Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: Ica7efcb77d760652e76718d72a3be38b43b35c50