IB/hfi: Fix user SDMA packet size for smaller MTUs
User SDMA code calculates the size of the packet based
on the MTU size and expected RcvArray entry size, taking
the smaller of the two sizes - if the expected RcvArray
entry has enough room for entire MTU, the MTU is used,
otherwise, the room left in the RcvArray entry size.

The previous code worked well for MTUs of 4K or larger
due to an optimization in the code. When calculating the
size of the packet, only the second packet was checked
against the RcvArray entry size. This worked because with
that size MTU, two packets would cover a TID pair and the
offset into the RcvArray entry would get reset to 0.

For MTUs smaller than 4K, more than two packets can fit in
a TID Pair so the optimization check above failed to calculate
the correct size at the end of the TID pair.

This change fixes that behavior by extending the optimization
check to the number of packets that could fit into a 8K TID
pair.

INTERNAL: PR 127491 - 2K mtu broken in SDMA mode

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I9ef3603c6a0e91cc2287a071078ade4c1284ec63