IB/hfi: Fix usage of per-VL MTUs for Verbs
Unlike InfiniBand, the STL architecture supports
per-VL MTUs, which are set by the Fabric Manager.

Due to this difference, we should not be using the
port MTU as the setting for the QP MTU values and
also should not be using it to sanity check the
length of UD packets.

This commit changes handling of QP MTU by ensuring
that RC and UC QP MTUs are set according to the
MTU of the proper VL and UD QPs return an error if
the size of the work request is larger than the MTU
of the VL.

This is done under the covers of the Verbs layer. In
the case of RC and UC, this does not matter as everything
still work and the user will never know. However, for
UD this does create confusion due to the fact that an
error could be returned for smaller packets than the
reported port MTU.

INTERNAL: PR 125235 - ib_send_bw is unable to pass message
INTERNAL: sizes larger than the MTU configured for a VL
INTERNAL: when the VL's MTU is 2048B or less

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: Ib593e4b46485f2aa23dda8baaeb5d914632462c6