IB/hfi: add in-tree sync scripts and revise makefile
The scripts are suitable for use in moving the driver
into wfr-linux-devel.

The scripts assume that the initial driver for-<branch>
has been created and prepared as follows:

1. Checking out the driver with the for-<branch> from master
2. Determining the driver version with "make version"
3. Removing the rdma directory and the makefile
4. Editing common.h with the driver version
5. commiting and pushing the changes to for-<branch>

Part of these change change make version to just report the
version.   This makes the sync much easier

INTERNAL: PR 127362 - Be able to build WFR driver in tree

Change-Id: I20d1f1964a2544a2866981ec76821ce4a086d4da
Reviewed-by: Jubin John <jubin.john@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>