IB/hfi: implement 'SMA Packet Checks' per STL spec
STLv1, rev 0.90, section 9.10.26 specifies checks on incoming SMPs.
This patch implements those checks for both "local", and "non-local"
SMPs. A "local SMP" is one that originates from, and is destined to
the same node, i.e., it is looped back (in s/w). A "non-local" SMP
is one that arrived 'on the wire' from a different node.

While I'm here, did a little refactoring of code in qib_ud_rcv() -
specifically, change things so we read the SMPs pkey in only one
place.

INTERNAL: PR 127294 - wfr is too liberal when accepting for limited
INTERNAL: management pkey

Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: I5c8d001527430ab53ab76db01396202593f5b640