IB/hfi: Add 31 bit PSN for verbs
Use delta_psn() and cmp_psn() for all PSN comparisions.  This is
required to switch to a 31 bit verbs PSN.  The nature of the
routines switches from 24 bit to 31 bit based on
CONFIG_HFI1_VERBS_31BIT_PSN.

Additionally, the msn compare has been changed from qib_cmp24()
to cmp_msn() for clarity and to git rid of a qib'ism.

A Kconfig/CONFIG_HFI1_VERBS_31BIT_PSN configuration variable
to is used to enable/disable.   The default is enabled based on
current makefile and Kconfig.

INTERNAL: PR 124709 - Change WFR driver to use 31-bit PSN for Verbs

Change-Id: I2d97560ccf27b69a5c36ddf8dee70f347f50e6b8
Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>