IB/hfi: fix hfi_cdbg() panic with RHEL7 native kernel
The root cause is due to a compiler optimization for the hfi_cdbg() with the
RHEL7 config file together with having strings loaded into read only pages.

There is an inline routine that checks for a newline at the end for the format
routine to get rid of gratuitous newlines at the end of the format string by
writing a '\0' at the end of the string.

The compiler optimizer sees the newline, gets rid of the strlen and len tests
and optimizes the code to do a movb into the (read only) string resulting in
the panic when the caller HAS a newline at the end of the format string.

The fix is two fold:

1. Get rid of the writing of the '\0'
2. Get rid of hfi_cdbg() newline use cases

INTERNAL: PR 128075 - Panic loading hfi driver on native RHEL7.0 kernel

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>