IB/hfi1: Fix CableInfo SMA queries beyond byte 511
When a CableInfo SMA query is issued to a WFR port for CableInfo
bytes 512-4095, the SMA returns bad status. This was due to an
error in the count of the maximum number of 128 byte QSFP pages and a
limitation in the get_cable_info(...) function. The get_cable_info(...)
function returned the entire data buffer set to zero even if only part of
the address range requested was beyond the maximum memory available
on a QSFP cable per SFF 8636. This patch fixes the limitation and
returns valid data for the valid part of the requested address range.

INTERNAL: PR 128041 - WFR SMA returns bad status for CableInfo
INTERNAL: queries to addresses >511B

Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Signed-off-by: Easwar Hariharan <easwar.hariharan@intel.com>
Change-Id: If2f5313ab42806d090bcbd893818578c26a1a00b