IB/hfi1: more checkpatch fixes
This eliminates an error in wfr.c and a warning in diag.c.

The OVERFLOW_ELEM() macro needed to be split up.   To mitigate the
additional lines of code, macros were used to reduce the LOC in
wfr.c resulting in a new LOC reduction.

The diag.c had a return in macro.  That is resolved by changing
the macro to a routine.   Since the code path isn't performance
critical, that will be ok.

A similar WARNING exists in the PSM/sdma code for AHG, but
that is in a performance path, so that isn't changed.

At this point we are left with C99 errors, volatile stuff,
and unneeded memory error message warnings.

The C99 errors are picked in the another PR as part of the
FIXME resolution.

The remaining effort here is to insure with a test that no
additional checkpatch issues creep in.

INTERNAL: PR 124196 - checkpatch fixes as required

Change-Id: I1110b74eeb51dcf54811e52bb1d6a875641e9b7f
Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>