IB/hfi1/tests: add checkpatch build test
This patch adds a checkpatch to test/tests/build.sh to
detect a difference between the current source and a
checkpatch.whitelist.

The whitelist will need to be maintained as additional failures
are eliminated.

A failure of the whitelist diff will be detected by
the log grep and email.   Nothing will fail about
the nightly tests.

INTERNAL: PR 124196 - checkpatch fixes as required

Change-Id: I33ba6b615ec30db04e88021ee1ec9b7e3feea42e
Reviewed-by: Jubin John <jubin.john@intel.com>
Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>