Add TX counting events to kernel client POC
Thanks to Brian for pointing out that TX CT enable flag is required
for the md options in the TX command after which TX counting events
work successfully. The test is expected to pass only if firmware is
enabled but it should not hang if firmware is not enabled and fail
with an error instead.

Cc: Sanath Kumar <sanath.s.kumar@intel.com>
Reviewed-by: Brian Welty <brian.welty@intel.com>
Signed-off-by: Sudeep Dutt <sudeep.dutt@intel.com>