Ib/hfi1: Remove not applicable comments.
Diag.c fixme was tagged back when snoop first went in and wfr-lite was being
used as a model. It had extra things which are not needed on STL links. This
comment can just be removed.

The mad.c fixmes was added before counters were fully supported. I don't think
this is applicable anymore.

Remove comment about a per VL counter not implemented  in wfr.c as this is
not required. Also removes the TODO for converting the guid to host byte order
and only switching to network when we stuff it into a packet. This is a nice
to have but very low priority, filed PR 128212 to address at some point.

INTERNAL: PR 127355 - Remove all TODO and FIXMEs

Change-Id: Ibc204033c14fe74c2e64274c4375f28a8c2c0444
Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>