IB/hfi1: Remove loadtime_init FIXME
The commented out code in loadtime_init() does not apply to WFR.
Remove both it and the FIXME.

Deliberately leave the function loadtime_init() empty.
It will either be added to or removed when support for
driver re-init is added or removed.

INTERNAL: PR 128209 - Add or remove support for driver re-init

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>