IB/hfi1: Implement DC power saving based on cable presence
There is no reason for the blocks in the DC to be running
if there is no cable plugged into the system or the user
has explicitly disabled the link on the port.

Therefore, we can turn the block off (put them in reset) to
save on power usage until a cable is plugged in or the user
re-enables the link.

When the cable is re-inserted, the QSFP interrupt will
notify us and we'll turn the DC blocks back on prior to
starting to go through the link state machine.

If the DC was shut down due to user request, the DC will
be brought out of reset when the link is switched to POLL
state.

INTERNAL: PR 124744 - Understand and implement power and
INTERNAL: thermal management for WFR in driver

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I113edba79b7ccb4780e0a96144593221e065b747