IB/hfi: turn on HFI_CAP_STATIC_RATE_CTRL by default
Until now we've not turned on HFI_CAP_STATIC_RATE_CTRL by default.
As a result, PSM cannot use CCA (for example) unless that capability
is explicitly toggled on. This patch changes the default capability
mask to enable HFI_CAP_STATIC_RATE_CTRL (for user, and kernel).

In addition, change the function to display the 'hfi_cap_mask',
so that it reports its result in hexadecimal.

INTERNAL: PR 127801 - Static rate control needs to be enabled by
INTERNAL: default within driver and PSM

Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>