IB/hfi: cache VL Arb Tables for efficiency, and spec compliance
The STLv1 spec says that unused elements of the VL arbitration
table should contain VL15, and a weight of 0. However, WFR hardware
supports only 3 bits for the VL - only enough for VLs 0 through 7.
To compensate, we'll cache the VL arbitration tables, and return
the cached data (including the 4th bit of the VLs).

In a Set(VLArbTable) we'll write to the relevant CSRs only if
there is a change relative to the cached values. In a
Get(VLArbTable) we'll always return the cached values, and
never read CSRs to get the data. This is obviously the correct
thing to do after any Set(VLArbTable). But it's also correct
prior to any Set(VLArbTable), since then both the cache, and the
relevant CSRs will be zeroed.

This has an additional benefit in that we will no longer exercise
the (potentially disruptive) "drain VL" code in set_vl_weights()
unless there is an update to the VL arbitration tables that
requires modifying the CSRs.

INTERNAL: PR 128138 - VLARB Table SETS do not echo back the proper data

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: I73912fdf6acac41fb04b9fdf4c15692cc899f97f