IB/hfi1: Remove subvendor restore (DBI registers)
DBI is not a green register, so remove all references to
it.  This means that the workaround for the A0 erratum that
erases subvendor and subvendorid on an SBR is removed.  All A0
boards will not be able to retain a different subvendor value
if the driver is used to bring PCIe link speed to Gen3.

INTERNAL: PR 127355 - Remove all TODO and FIXMEs

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>