Ib/hfi1: Clean register accesses
This patch removes the WFR tag from registers as well as restricts which
registers are included in the now single header file to just those used by the
driver.

All other references to WFR will be removed in the follow on patch. It is
being separated out to make rebase and code review easier.

INTERNAL: PR 127340 - Remove "wfr" from driver source base.

Change-Id: Ifc2a6a4acefb835dd3957589d6fb78f75c907690
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>