Ib/hfi1: Remove diag device file and associated observer code
As far as we can tell this device file is totally unused these days. It was
used in qib for debugging but not with hfi. hfidiags uses the ui device, snoop
used hfi1_pktX and hfi1_diagpkt_send uses hfi1_diagpkt. This patch removes all
remants of the diag device and observer code.

INTERNAL: PR 127492 - Remove diag observer code and associated logic in diag.c

Change-Id: Iad683645049da905874dd237baefcb2feab5c66e
Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>