IB/hfi1: Add look-aside cache constructor function
The kernel slub allocator attempts to merge memory slabs
of the same sizes together for efficiancy. This manifests
itself during the creation of the SDMA request look-aside
cache.

Unfortunately, there seems to be an issue with the slub
allocation code in that mergable slabs are assigned a name
in the sysfs directory struct which is generated by the
kernel.

There is one look-aside cache for each process and since all
caches are of the same size, the kernel should be merging
them together. However, it doesn't appear to be doing so,
which creates a name collision with the kernel-generated name.

The workaround is rather simple and should work on all
kernel versions - provide a look-aside cache constructor
function, which can be used to memset the objects to 0.

The constructor is somewhat useless since object initialization
still needs to be done when the object is allocated (in case
of object reuse), but the constructor is needed and an empty
contructor is even more useless.

INTERNAL: PR 128306 - mvapich2 PSM - IFS 403 - mpi apps fails
INTERNAL: with > -np 2 - Cannot allocate memory

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I8add8169fdbfb8f3e3e212fe10f4d336cc49081a