libnd: stable(ish) block device names
From Toshi:

    We have seen a case that the region#->pmem# binding becomes
    inconsistent across a reboot when there are 8 NVDIMM cards
    (reported by Robert Elliott).  This leads user to access a
    wrong device.

While it is a bug for userspace to depend on stable device names from
boot-to-boot, we should try not to give random results in the case where
the configuration has not changed.  Now that PMEM only ever attaches to
'region' devices emitted by libnd, we can simply carry the region-id all
the way down to the pmem device name.

Since BLK-regions allow multiple namespaces per-region we need to carry
the <region-id>.<namespace-index> in the block device name if we want
stable(ish) device names.

Cc: Christoph Hellwig <hch@lst.de>
Reported-by: Toshi Kani <toshi.kani@hp.com>
Reported-by: Robert Elliott <Elliott@hp.com>
Reviewed-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>