IB/hfi1: Break HW mutex if held too long
We have had a case where one HFI on a dual port card went down
holding the hardware mutex.  This blocked the other HFI.  Change
the acquire to break the mutex if it is held too long.

INTERNAL: PR 128444 - Cannot drop the mutex once the PCIe interface
INTERNAL: is gone

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>