IB/hfi: fix problems introduced in commit dce1b00
Commit dce1b00 caused the WFR_SEND_{CTXT,DMA}_CHECK_ENABLE registers
to be set differently depending on the "HFI_CAP*" flags (specifically
the HFI_CAP_PKEY_CHECK, and HFI_CAP_STATIC_RATE_CTRL flags). This
is not always what we want - for example, it generally doesn't make
sense to enable 'CheckPartitionKey' for an SDMA engine, since an
SDMA engine can send on any partition, and the egress check is
against a single partition key.

This patch removes adjustments to the WFR_SEND_{CTXT,DMA}_CHECK_ENABLE
registers that were made when the context is created, and instead
just uses the defaults to initialize the registers at context creation
time. Only if a partition key is set, is the HFI_CAP_PKEY_CHECK
capability considered (and even then, only for user contexts). The
existing hfi1_init_ctxt() code is used to toggle
'DisallowPbcStaticRateControl' after a context is created depending
on the value of the HFI_CAP_STATIC_RATE_CTRL.

In addition:
  - HFI_CAP_PKEY_CHECK << HFI_CAP_USER_SHIFT was added to
    HFI_CAP_MASK_DEFAULT. This means that the value reported
    for this bit in the corresponding sysfs file now matches
    what the driver sees when testing HFI_CAP_IS_USET(PKEY_CHECK).
  - The 'restore' parameter was removed from adjust_integrity_checks().
    It's not needed anymore, since it can be inferred from the
    'hfi_snoop.mode_flag'.

INTERNAL: PR 127987 - driver cannot enable static rate control

Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: I501170ab3ca85bf77fa5b7df5eadc2292559c907