IB/hfi1: Remove TBD errata
There are two TBD errata implemented in the driver, 291496 and 291500.
According to the errata spreadsheet, code/comments need to be cleaned up
of references to TBD erratum.

This commit removes the comments and prints which refer to these two
TBD errata.

INTERNAL: PR 127341 - Remove all non public info from the driver

Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Jubin John <jubin.john@intel.com>
Change-Id: I744570000ed2f03d8e06f2f954ae7d75b331b909