IB/hfi1: Remove fw_validate
Remove the ability to turn firmware validation on or off.
This has two goals for release outside Intel:

o Remove the idea that this is optional.  It will not be
  optional in the finished product.
o Obfuscate the workaround needed for A0 to flip firmware
  validation

Enhance the obfuscation by renaming the disable validation
bit to "enable spico".

INTERNAL: PR 128268 - Open source software will not contain
INTERNAL: workarounds for red errata since red features must
INTERNAL: not be exposed in such software

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>