IB/hfi1: Fix SnoopIoctl issues
Fix issues found in the driver when running the SnoopIoctl test:
o When moving from disabled to offline, turn on the dc.  This
  pathway was missed.
o When going to disabled, extend the check that decides whether to
  call the link down function.  Add LINK_COOLDOWN and disabled as
  new reasons not call the link down function.
o When shutting down the dc, only reset the 8051 with DC8051_CFG_RST.
  Leave CRAM, DRAM, IRAM, and SFR alone.

In addition:
o Add a print if the 8051 times out restarting in dc_start().
o Extend the wait for the 8051 to start in goto_offline() to 3s.

INTERNAL: PR 128583 - Intermittent SnoopIoctl failure possibly
INTERNAL: due to 8051 command timeout

Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>