IB/hfi1: Proper locking for thermal initialization
The thermal sensor needs to be initialized once per ASIC
rather than once per device. This requires that the driver
figure out which of the available, initialized devices is
on the same package.

This allows the driver to initialize the thermal sensors on
each chip in the host but only once per chip.

This commit also implements a workaround in response to a HW
behavior which could result in a lost
ASIC_STS_SBUS_RESULT.RCV_DATA_VALID bit. The workaround
requires that in the case of a timeout waiting for the bit to
go high, the driver check the SBus counters to ensure that the
command has succeeded.

INTERNAL: PR 128574 - therm_init() should run once for the ASIC

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: Iddc8671e5b4368076e06b91ef51087a54d3c52b3