IB/hfi1: Convert trace to use trace_seq_buffer_ptr helper
This patch incorporates upstream commit 7b039cb4c5a9
("tracing: Add trace_seq_buffer_ptr() helper function")
For the original description use 'git show 7b039cb' in a
full kernel source tree

This patch has been modified from the original as necessary
to support the hfi1 driver. This patch is part of a series to
move the driver to the latest upstream code base and cannot be
applied to master because an upstream commit this patch
depends on is in a kernel newer than 3.12.18.

INTERNAL: PR 127361 - Update driver to the latest upstream code base

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Jubin John <jubin.john@intel.com>
Change-Id: Id3ea204cd4944c0ffab7ccf435a67c435c0d757e