IB/hfi1: Add TRACE_SYSTEM_VAR
This patch is derived from upstream commit 882156e0407f
("tracing: Add TRACE_SYSTEM_VAR to intel-sst"). For the original
description use 'git show 882156e' in a full kernel source tree.

This patch has been modified from the original as necessary
to support the wfr driver.This patch is part of a series to
move the driver to the latest upstream code base and cannot be
applied to master because an upstream commit this patch
depends on is in a kernel newer than 3.12.18.

This patch is needed because of upstream commit acd388fd3af3
("tracing: Give system name a pointer")

INTERNAL: PR 127361 - Update driver to the latest upstream code base

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Jubin John <jubin.john@intel.com>
Change-Id: I3350a3c054c72518438c67af9d5621452e8130d7