IB/hfi1: Use pcie_enable_msix_range()
This patch incorporates qib upstream commit bf3f043e7bc2
("IB/qib: Use pci_enable_msix_range() instead of pci_enable_msix()")
For the original description, use 'git show bf3f043'
in a full source tree

This patch has been modified from the original as necessary
to support the hfi1 driver. This patch is part of a series to
move the driver to the latest upstream code base and cannot be
applied to master because an upstream commit this patch
depends on is in a kernel newer than 3.12.18.

INTERNAL: 127361 - Update driver to the latest upstream code base

Reviewed-by: Dean Luick <dean.luick@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Jubin John <jubin.john@intel.com>
Change-Id: I3ba5750db910e271898c610c9e7c2bd8210700f2