opa_core interface for SMA
Since SMA has combination of HW and SW attribute components to be
set or queried, it is broken down into two entities
1. HW entity which resides on the opa*_hfi driver
2. SW/verbs entity which resides on the opa_ib driver

A SMA MAD packet is thus processed by these two entities
with first preference given to the HW entity.
This way for the future gen products, if a particular
HW attribute method is not supported then that method can
be easily rejected without having to change the opa_ib
driver.

Change-Id: I69b4b105e6ae2befcc3f154d5f04e6716bef3360
Reviewed-by: Brian Welty <brian.welty@intel.com>
Signed-off-by: Sanath Kumar <sanath.s.kumar@intel.com>