IB/hfi1: mask ULP passed PSNs to 24 bits
It looks like some sloppy ULPs pass PSNs with
data in the upper 8 bits of the 32 bit container
passed in the QP modify.  One of those sloppy
ULPs is rdma_cm.

The 31 bit PSN logic is patched to mask PSN sets
from a ULP to 24 bits.  Queries of a PSN are left
to return the full 31 bit value in the 32 bit
container.

INTERNAL: PR 128335 - rdma_cm not working on upstream kernels

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>