IB/hfi1: Set PCIe parent target to Gen3 if needed
Our SuperMicro test systems have a BIOS that sets the parent
PCIe link target to Gen2 even though it is Gen3 capable.  Add
code to move the parent the target speed if at a lesser speed.
Note: The driver has already checked if the parent is Gen3 capable
so that is not checked here.

Only write to the parent if needed.

INTERNAL: PR 128690 - HFI card coming up at Gen2 speed in
INTERNAL: SuperMicro server

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>
Change-Id: I084379d0602d833f61baa742e041e693874bfc6a