IB/hfi1: Do not enable thermal polling across pcie bump
Thermal polling being enabled is bad news when trying to transition to pcie
gen 3. This patch disables thermal polling if it is enabled as part of the
pcie gen 3 transition. There is also an apparent bug in the Sbus master
firmware that requires the Sbus master to restart from a clean slate each time
so enable the firmware download on each pcie gen 3 transition.

INTERNAL: PR 128711 - Thermal polling is enabled across PCIe firmware swap
INTERNAL: leading to lost device

Change-Id: Ic3ca525b4eebd2dbead4ef16172d0601726ba573
Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>