nfit: Fix up address spaces, sparse warnings
Fix a couple of sparse warnings (duplicate const, incorrect address
space), and go back to using memcpy() instead of memcpy_toio() and
memcpy_fromio() when talking to our block apertures.  Instead, include a
union to alias mmio->base and mmio->aperture so that we can continue to
reuse common code for ioremapping and deinterleaving.  mmio->base still
has the __iomem annotation and is used via readq() and writeq() for the
control and status registers.  mmio->aperture is used via normal
memcpy() for aperture I/O.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>