IB/hfi: fix sparse warnings for verbs AHG
sdma_build_ahg_descriptor() accepts a u16 for
the value to use in an edit.

The endianness of that u16 can vary depending on the
endianness of the part of the header that is to be
edited and cannot be determined from the common routine.

To preserve existing functionality, just use force
to silence the errors.

The patch resolves these sparse issues:

ruc.c:751:42: warning: cast from restricted __be16
ruc.c:760:50: warning: cast from restricted __be16

INTERNAL: PR 128771 - Sparse endianness checks and fixes

Change-Id: I0cffa227aecf950cf0226734257f1e9e719902a7
Reviewed-by: Harish Chegondi <harish.chegondi@intel.com>
Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>