IB/hfi: handle verbs static rates
Up to now the static rate has been ignored when sending a packet
using verbs. This patch takes the QP's static rate into account, and
adds a delay value to the PBC, if necessary. The context, or SDMA
engine must be configured to allow per-PBC static rate control for
this to be effective, but the default is to allow per-PBC static
rate control.

The patch also includes some refactoring of code calculate rates,
and delays, since these are also used for Congestion Control.

Also found that hfi1_init_ctxt() was not being called for the
per-VL send contexts. Since hfi1_init_ctxt() sets the flags for
egress checks, including per-PBC static rate control, that was
changed.

INTERNAL: PR 128554 - static rate from verbs ULP not handed

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: Iffa7e35bbe19488c572745e50d0d3130608830de